+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.6-pre
 * Add error status to differentiate error from check failures
 * Add success, failure, and error messages to improve reporting
+* Rename responseTime to duration (check time in milliseconds)
 
 ### 0.0.5
 * Add cli formatter